Fix printMuzzleReferences gradle task
